+++ Examples/Package.swift
@@ -10,6 +10,7 @@ targets.append(executable("router"))
 targets.append(executable("dsl"))
 targets.append(executable("echoSlashCommand"))
 targets.append(executable("threadExpander"))
+targets.append(executable("blockActionsMessageContainer"))
 
 let package = Package(
     name: "Examples",