+++ Sources/SlackClient/Interactions/Payloads/BlockActionsPaylaod.swift
@@ -14,7 +14,7 @@ public struct BlockActionsPaylaod: InteractivePayloadProtocol, Decodable, Sendab
     public let apiAppId: String?
     public let actions: [ActionElementType]?
     public let channel: Channel?
-    public let view: View
+    public let view: View?
 
     private enum CodingKeys: String, CodingKey {
         case _type = "type"
@@ -31,6 +31,6 @@ public struct BlockActionsPaylaod: InteractivePayloadProtocol, Decodable, Sendab
 
 extension BlockActionsPaylaod {
     public var callbackId: String? {
-        view.callbackId
+        view?.callbackId
     }
 }